Introduce small step delays
Rather than delaying in switch statement for each STATE, make each
"state" a function that returns whether or not it is currently
"running" based on elapsed time. If a state function returns 0, call a
function to set the next state and reset the elapsed timer. End with
stopped bot, ready for teleop.